main3

+++ 05/conversion/main.go
@@ -6,6 +6,7 @@ func main() {
 	speed := 100
 	force := 2.5
 
-	speed = speed * int(force)
+	// speed = speed * int(force)
+	speed = float64(speed) * force
 	fmt.Println(speed)
 }
